doc-tools/requirements.txt
beautifulsoup4>=4.9.3
requests>=2.25.1
python-dotenv>=0.19.0
openai>=1.0.0
crawl4ai>=0.4.0
playwright>=1.30.0
aiohttp>=3.8.1
typing-extensions>=4.0.0
tenacity>=8.2.0
# New dependencies for post-processor
pyyaml>=6.0
numpy>=1.21.0
scikit-learn>=1.0.0
networkx>=2.6.0
